Fix package deps
