--- charts/child-state-update-provider/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@lg-charts/child-state-update-provider",
-  "version": "0.0.0",
-  "description": "lg-charts ChildStateUpdateProvider Component",
-  "main": "./dist/index.js",
-  "module": "./dist/esm/index.js",
-  "types": "./dist/index.d.ts",
-  "license": "Apache-2.0",
-  "scripts": {
-    "build": "lg build-package",
-    "tsc": "lg build-ts",
-    "docs": "lg build-tsdoc"
-  },
-  "publishConfig": {
-    "access": "restricted"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/mongodb/leafygreen-ui"
-  },
-  "bugs": {
-    "url": "https://jira.mongodb.org/projects/PD/summary"
-  }
-}
