--- charts/child-state-update-provider/src/ChildStateUpdateProvider/ChildStateUpdateProvider.types.ts
@@ -1,5 +1,0 @@
-import { PropsWithChildren } from 'react';
-
-export type DragProviderProps = PropsWithChildren<{
-  onDragEnd?(event: { active: string; over: string }): void;
-}>;
