--- charts/child-state-update-provider/src/index.ts
@@ -1,4 +1,0 @@
-export {
-  ChildStateUpdateProvider,
-  useChildStateUpdateContext,
-} from './ChildStateUpdateProvider';
