--- charts/child-state-update-provider/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "extends": "@lg-tools/build/config/package.tsconfig.json",
-  "compilerOptions": {
-    "declarationDir": "dist",
-    "outDir": "dist",
-    "rootDir": "src",
-    "baseUrl": ".",
-    "paths": {
-      "@leafygreen-ui/icon/dist/*": ["../../packages/icon/src/generated/*"],
-      "@leafygreen-ui/*": ["../../packages/*/src"]
-    }
-  },
-  "include": [
-    "src/**/*"
-  ],
-  "exclude": ["**/*.spec.*", "**/*.stories.*"],
-  "references": [
-  ]
-}
