--- charts/drag-provider/src/DndKitProvider/index.ts
@@ -1,1 +1,0 @@
-export { DndKitProvider } from './DndKitProvider';
